Update stubs.md
It would be helpful for `python3 -m deal stub /path/to/a/file.py` to detect whether `file.py` already has deal annotations and generate itself based on them. This could be a neat way to separate deal contracts into their own file rather than intrude on the code, sort of like it's possible to do with type hints.